Fix publish settings
