+++ .github/workflows/ci.yml
@@ -65,12 +65,9 @@ jobs:
       - name: Check that workflows are up to date
         run: sbt '++${{ matrix.scala }}' 'project /' githubWorkflowCheck
 
-      - name: Check that common scalafmt config is up to date
-        run: sbt '++${{ matrix.scala }}' 'project /' lucumaScalafmtCheck
-
       - name: Check headers and formatting
         if: matrix.java == 'temurin@8'
-        run: sbt '++${{ matrix.scala }}' headerCheckAll scalafmtCheckAll 'project /' scalafmtSbtCheck
+        run: sbt '++${{ matrix.scala }}' headerCheckAll scalafmtCheckAll 'project /' scalafmtSbtCheck lucumaScalafmtCheck
 
       - name: Test
         run: sbt '++${{ matrix.scala }}' test
